Avoid useless position values writing and allowing "third party" inputs
Now the hardware component has a parameter that, if set to false sends
the position commands to the control board wrapper only if they are
different from the ones previously sent. Another feature enabled by
putting this flag to false is the possibility to still control the robot
joints from yarp, even if they are "claimed" by the ros2 controller.